+++ js/components/dataset/resource/form.vue
@@ -58,40 +58,105 @@
         }
     }
 }
+
+.resource-upload-dropzone {
+    border: 4px dashed @gray-lte;
+    min-height: 150px;
+    padding-top: 10px;
+    padding-bottom: 10px;
+    margin-bottom: 15px;
+
+    .lead {
+        margin-bottom: 0;
+    }
+
+    .drop-active > & {
+        border: 4px dashed @green;
+    }
+
+    // url field
+    .col-sm-9 > div {
+        margin-right: 15px;
+    }
+}
+
+.resource-choose-upload {
+    text-align: center;
+    a {
+        text-decoration: underline;
+    }
+}
+
+.modal-body {
+    .resource-choose-upload {
+        a {
+            color: white;
+        }
+    }
+}
+
 </style>
 
 <template>
 <div>
-<div class="list-group actions-list" v-if="!resource.filetype">
-    <a v-for="action in actions" class="list-group-item pointer"
-        @click="set_filetype(action.filetype)">
-        <div class="action-icon">
-            <span class="fa fa-3x fa-{{action.icon}}"></span>
+    <form-horizontal v-if="resource.url || hasChosenRemoteFile" class="resource-form file-resource-form"
+        :fields="fields" :model="resource" v-ref:form>
+    </form-horizontal>
+    <div v-show="files.length" v-for="file in files" class="info-box bg-aqua">
+        <span class="info-box-icon">
+            <span class="fa fa-cloud-upload"></span>
+        </span>
+        <div class="info-box-content">
+            <span class="info-box-text">{{file.name}}</span>
+            <span class="info-box-number">{{file.size | filesize}}</span>
+            <div class="progress">
+                <div class="progress-bar" :style="{width: progress+'%'}"></div>
+            </div>
+            <span class="progress-description">
+            {{progress}}%
+            </span>
         </div>
-        <h4 class="list-group-item-heading">
-            {{ action.label }}
-        </h4>
-         <p class="list-group-item-text ellipsis">
-        {{ action.details }}
-        </p>
-    </a>
-</div>
-<component v-ref:form v-if="resource.filetype" :is="form"
-    :dataset="dataset" :resource="resource">
-</component>
+    </div>
+    <div class="resource-choose-upload" v-if="showUploadZone">
+        <div class="resource-upload-dropzone">
+            <div class="row">
+                <div class="text-center col-xs-12">
+                    <span class="fa fa-download fa-4x dropicon"></span>
+                </div>
+                <div class="text-center col-xs-12 lead">
+                    <p>{{ _("Drag a file here") }}</p>
+                </div>
+                <div class="text-center col-xs-12">
+                    <p>{{ _( "or") }}</p>
+                </div>
+                <div class="text-center col-xs-12">
+                    <span v-el:upload-btn class="btn btn-outline btn-flat">
+                    {{ _("Select a file from your computer") }}
+                    </span>
+                </div>
+            </div>
+        </div>
+        <div v-if="!hasUploadedFile">
+            <a href="" @click.prevent.stop="hasChosenRemoteFile = true">
+                {{ _("You can also link to an existing remote file or URL by clicking here.") }}
+            </a>
+        </div>
+    </div>
 </div>
 </template>
 
 <script>
+import API from 'api';
 import Dataset from 'models/dataset';
 import Resource from 'models/resource';
-
-import FileForm from 'components/dataset/resource/file-form.vue';
-import RemoteForm from 'components/dataset/resource/remote-form.vue';
-import ApiForm from 'components/dataset/resource/api-form.vue';
+import CommunityResource from 'models/communityresource';
+import FormHorizontal from 'components/form/horizontal-form.vue';
+import UploaderMixin from 'mixins/uploader';
+import resource_types from 'models/resource_types';
 
 export default {
-    components: {FileForm, RemoteForm, ApiForm},
+    components: {FormHorizontal},
+    mixins: [UploaderMixin],
     props: {
         dataset: {
             type: Object,
@@ -99,38 +164,128 @@ export default {
         },
         resource: {
             type: Object,
-            default() {return new Resource()}
+            default() {
+                return new Resource({
+                    data: {
+                        filetype: 'remote',
+                        type: 'main',
+                    }
+                }
+            )}
         },
-        hideNotifications: false
+        hideNotifications: false,
     },
     data() {
         return {
-            actions: [{
-                label: this._('Local file'),
-                details: this._('Send a file from your computer'),
-                icon: 'cloud-upload',
-                filetype: 'file'
-            }, {
-                label: this._('Online file'),
-                details: this._('Register an already online file'),
-                icon: 'cloud',
-                filetype: 'remote'
-            }, {
-                label: this._('API'),
-                details: this._('Register an API to access data'),
-                icon: 'puzzle-piece',
-                filetype: 'api'
-            }]
+            hasChosenRemoteFile: false,
+            generic_fields: [{
+                    id: 'title',
+                    label: this._('Title')
+                },{
+                    id: 'type',
+                    label: this._('Type'),
+                    widget: 'select-input',
+                    values: resource_types,
+                    map: function(item) {
+                        return {value: item.id, text: item.label};
+                    }
+                }, {
+                    id: 'description',
+                    label: this._('Description'),
+                }, {
+                    id: 'published',
+                    label: this._('Publication date'),
+                    widget: 'date-picker'
+                }],
+            file_fields: [{
+                    id: 'url',
+                    label: this._('URL'),
+                    readonly: false,
+                    widget: 'url-field',
+                },{
+                    id: 'filesize',
+                    label: this._('Size'),
+                    readonly: false
+                }, {
+                    id: 'format',
+                    label: this._('Format'),
+                    widget: 'format-completer',
+                    readonly: false
+                }, {
+                    id: 'mime',
+                    label: this._('Mime Type'),
+                    readonly: false
+                }, {
+                    id: 'checksum',
+                    label: this._('Checksum'),
+                    widget: 'checksum',
+                    readonly: false
+                }],
+            upload_multiple: false,
+            progress: 0,
         };
     },
     computed: {
-        form() {
-            return `${this.resource.filetype}-form`;
-        }
+        showUploadZone() {
+            if (this.files.length) return false;
+            if (this.hasChosenRemoteFile) return false;
+            if (this.resource.url && !this.hasUploadedFile) return false;
+            return true;
+        },
+        hasUploadedFile() {
+            return this.resource.filetype == 'file';
+        },
+        fields() {
+            return this.generic_fields.concat(this.file_fields);
+        },
+        is_community() {
+            return this.resource instanceof CommunityResource;
+        },
+        upload_endpoint() {
+            const operations = API.datasets.operations;
+            let params = {};
+            if (typeof this.dataset !== 'undefined') {
+                params = {dataset: this.dataset.id};
+            }
+            if (this.resource.id) {
+                if (this.is_community) {
+                    params.community = this.resource.id;
+                } else {
+                    params.rid = this.resource.id;
+                }
+            }
+            const route_new = this.resource.id ? '' : 'new_';
+            const route_namespace = this.is_community ? 'community_' : 'dataset_';
+            const endpoint = `upload_${route_new}${route_namespace}resource`;
+            return operations[endpoint].urlify(params);
+        },
+    },
+    events: {
+        'uploader:progress': function(id, uploaded, total) {
+            this.progress = Math.round(uploaded * 100 / total);
+        },
+        'uploader:complete': function(id, response) {
+            if (!this.is_community) {
+                this.dataset.resources.unshift(response);
+            }
+            // Do not override an existing typed or registered title.
+            if (this.$refs.form) {
+                const title = this.$refs.form.serialize().title || this.resource.title;
+                if (title) {
+                    response.title = title;
+                }
+            }
+            this.resource.on_fetched({obj: response});
+            this.postUpload();
+        },
     },
     methods: {
-        set_filetype(filetype) {
-            this.resource.filetype = filetype;
+        postUpload() {
+            this.resource.filetype = 'file';
+            this.file_fields = this.file_fields.map(ff => {
+                ff.readonly = true;
+                return ff;
+            });
         },
         serialize() {
             // Required because of readonly fields and filetype.
@@ -147,7 +302,7 @@ export default {
                 });
             }
             return isValid;
-        }
+        },
     }
 };
 </script>