+++ udata/core/dataset/api.py
@@ -43,12 +43,13 @@ from .api_fields import (
     frequency_fields,
     license_fields,
     resource_fields,
+    resource_type_fields,
     upload_fields,
 )
 from udata.linkchecker.checker import check_resource
 from .models import (
     Dataset, Resource, Checksum, License, UPDATE_FREQUENCIES,
-    CommunityResource
+    CommunityResource, RESOURCE_TYPES
 )
 from .permissions import DatasetEditPermission, ResourceEditPermission
 from .forms import (
@@ -508,3 +509,13 @@ class CheckDatasetResource(API, ResourceMixin):
         '''Checks that a resource's URL exists and returns metadata.'''
         resource = self.get_resource_or_404(dataset, rid)
         return check_resource(resource)
+
+
+@ns.route('/resource_types/', endpoint='resource_types')
+class ResourceTypesAPI(API):
+    @api.doc('resource_types')
+    @api.marshal_list_with(resource_type_fields)
+    def get(self):
+        '''List all resource types'''
+        return [{'id': id, 'label': label}
+                for id, label in RESOURCE_TYPES.items()]