+++ udata/frontend/helpers.py
@@ -345,7 +345,7 @@ def i18n_alternate_links():
                 url = url_for(request.endpoint, lang_code=lang, **params)
                 links.append(LINK_PATTERN.format(url=url, lang=lang))
         return Markup(''.join(links))
-    except:
+    except Exception:
         # Never fails
         return Markup('')