+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Current (in progress)
 
 - Normalize resource.format (migration - :warning: need reindexing) [#1563](https://github.com/opendatateam/udata/pull/1563)
+- Typed resources [#1398](https://github.com/opendatateam/udata/issues/1398)
 
 ## 1.3.5 (2018-04-03)